adds user.details()
