+++ packages/utils/src/global.js
@@ -1,0 +1,10 @@
+// GLOBAL
+getGlobalInterestGenerated(timePeriod) {
+  // TODO:
+  return {};
+}
+
+getGlobalInterestSent(timePeriod) {
+  // TODO:
+  return {};
+}
